Changed calling sequences for better optional parameter handling.
The uploadFile() bucketName paramter always had to be passed, and
setting it in the config has no effect.

Changed uploadFile() to remove the bucketName parameter. It uses the
config setting by default, and the caller can override the default by
specifying a `Bucket` setting in the `s3Options` parameter.

Changed getUploadURL() to move all parameters into an options parameter.